Restructure packaging for MBWay
COAND-497
