+++ mbway/src/main/java/com/adyen/checkout/mbway/internal/ui/MBWayDelegate.kt
@@ -6,7 +6,7 @@
  * Created by josephj on 13/7/2022.
  */
 
-package com.adyen.checkout.mbway
+package com.adyen.checkout.mbway.internal.ui
 
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
@@ -15,9 +15,11 @@ import com.adyen.checkout.components.ui.ButtonDelegate
 import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.components.util.CountryInfo
+import com.adyen.checkout.mbway.internal.ui.model.MBWayInputData
+import com.adyen.checkout.mbway.internal.ui.model.MBWayOutputData
 import kotlinx.coroutines.flow.Flow
 
-interface MBWayDelegate :
+internal interface MBWayDelegate :
     PaymentComponentDelegate<PaymentComponentState<MBWayPaymentMethod>>,
     ViewProvidingDelegate,
     ButtonDelegate,
